@@ -1,16 +1,14 @@
-﻿// Copyright (c) All contributors. All rights reserved.
+// Copyright (c) All contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using MessagePack;
 
 #pragma warning disable SA1307 // Accessible fields should begin with upper-case letter
 #pragma warning disable SA1401 // Fields should be private
+#nullable enable
 
-namespace TestData2
+namespace TestData2;
+
+[MessagePackObject(true)]
+public class C
 {
-    [MessagePackObject(true)]
-    public class C
-    {
-        public B b; public int a;
-    }
+    public B? b; public int a;
 }
